Refactor of src
