+++ FTestman/src/__init__.py
@@ -1,0 +1,4 @@
+from .module_requests import load_data_api, get_link, send_request
+
+__version__ = "0.1.0"
+__author__ = "AdamRix"
